+++ src/Nexus/src/Platform/D3D12/DeviceBufferD3D12.cpp
@@ -19,7 +19,7 @@ namespace Nexus::Graphics
 		// if the resource is a constant buffer, it needs to be aligned to 256 bytes
 		if (desc.Usage & Graphics::BufferUsage::Uniform)
 		{
-			m_BufferSize = Utils::AlignTo<size_t>(desc.SizeInBytes, 256);
+			m_BufferSize = Utils::AlignTo<uint64_t>(desc.SizeInBytes, 256);
 		}
 
 		D3D12_RESOURCE_DESC1 resourceDesc = {};
